moving some common files to lib
