+++ lib/electron-watcher.js
