+++ lib/exiftool.js
