+++ lib/root.js
