+++ renderer/filmstrip/rating.js
@@ -1,4 +1,4 @@
-const log = require('../../tools/log.js')('filmstrip-rating');
+const log = require('../../lib/log.js')('filmstrip-rating');
 const exiftool = require('../tools/exiftool-child.js');
 
 function ratingControl(rating) {
