+++ renderer/sidebar/sidebar.js
@@ -5,7 +5,7 @@ const name = 'sidebar';
 const style = fs.readFileSync(path.resolve(__dirname, `${name}.css`), 'utf8');
 
 const exiftool = require('../tools/exiftool-child.js');
-const log = require('../../tools/log.js')(name);
+const log = require('../../lib/log.js')(name);
 
 module.exports = function ({ events }) {
   var elem = document.createElement('div');
